predict_api/pyproject.toml
[tool.poetry]
name = "predict_api"
version = "0.1.0"
description = ""
authors = ["WxMAMADIANxW <mdjalo2@myges.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.21.1"
Pillow = "^8.3.1"
tqdm = "^4.61.2"
scikit-learn = "^0.24.2"
Flask = "^2.0.1"
gunicorn = "^20.1.0"
Flask-RESTful = "^0.3.9"
flask-restplus = "^0.13.0"
Werkzeug = "^2.0.1"
fastapi = "^0.67.0"
uvicorn = "^0.14.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
